add job duration, `WorkflowInvocationJob` component
remove `JobStepTabs.vue`
